Optimization class:ObjectTracking
change var type double to float
method:updateTracking 최적화
class:CameraTracking 한 traceline에서 하나의 점만 트래킹하도록 변경

+++ object-traking-ircamera/src/com/wei/cheapvr/Tracker/Number.java
@@ -3,10 +3,10 @@ package com.wei.cheapvr.Tracker;
 public interface Number<T> {
     T plus(T other);
     T minus(T other);
-    T mul(double r);
-    double distTo(T other);
+    T mul(float r);
+    float distTo(T other);
     
     
     String toString();
-    double abs();
+    float abs();
 }
