+++ object-traking-ircamera/src/com/wei/cheapvr/Tracker/Quaternion.java
@@ -4,15 +4,15 @@ import com.wei.cheapvr.Tracker.Quaternion;
 
 
 public class Quaternion implements Number<Quaternion> {
-    public double w, x, y, z;
+    public float w, x, y, z;
     
-    public Quaternion(double w, double x, double y, double z) {
+    public Quaternion(float w, float x, float y, float z) {
         this.w = w;
         this.x = x;
         this.y = y;
         this.z = z;
     }
-    public Quaternion(double w, Vector3 v) {
+    public Quaternion(float w, Vector3 v) {
         this.w = w;
         this.x = v.x;
         this.y = v.y;
@@ -24,8 +24,8 @@ public class Quaternion implements Number<Quaternion> {
         return new Quaternion(0, 0, 0, 0);
     }
     
-    public Quaternion mul(double w, double x, double y, double z) {
-        double newW, newX, newY, newZ;
+    public Quaternion mul(float w, float x, float y, float z) {
+        float newW, newX, newY, newZ;
         newW = this.w*w - this.x*x - this.y*y - this.z*z;
         newX = this.w*x + this.x*w + this.y*z - this.z*y;
         newY = this.w*y - this.x*z + this.y*w + this.z*x;
@@ -36,38 +36,38 @@ public class Quaternion implements Number<Quaternion> {
     public Quaternion mul(Quaternion q) {
         return mul(q.w, q.x, q.y, q.z);
     }
-    public Quaternion mul(double r) {
+    public Quaternion mul(float r) {
         return new Quaternion(w*r, x*r, y*r, z*r);
     }
-    public Quaternion plus(double w, double x, double y, double z) {
+    public Quaternion plus(float w, float x, float y, float z) {
         return new Quaternion(this.w + w, this.x + x, this.y + y, this.z + z);
     }
     public Quaternion plus(Quaternion q) {
         return plus(q.w, q.x, q.y, q.z);
     }
-    public Quaternion minus(double w, double x, double y, double z) {
+    public Quaternion minus(float w, float x, float y, float z) {
         return new Quaternion(this.w - w, this.x - x, this.y - y, this.z - z);
     }
     public Quaternion minus(Quaternion q) {
         return minus(q.w, q.x, q.y, q.z);
     }
-    public double distTo(Quaternion other) {
+    public float distTo(Quaternion other) {
         return sqrt(pow(abs(w - other.w), 2) + 
                         pow(abs(x - other.x), 2) + 
                         pow(abs(y - other.y), 2) + 
                         pow(abs(z - other.z), 2));
     }
     public Quaternion qProduct(Quaternion q) {
-        double w1 = this.w, w2 = q.w;
+        float w1 = this.w, w2 = q.w;
         Vector3 v1 = this.toVector(), v2 = q.toVector();
-        double w = w1*w2 - v1.dot(v2);
+        float w = w1*w2 - v1.dot(v2);
         Vector3 v = v2.mul(w1).plus(v1.mul(w2).plus(v1.cross(v2)));
         return new Quaternion(w, v);
     }
     
     
     
-    public static Quaternion getRollFromEuler(double rx, double ry, double rz) {
+    public static Quaternion getRollFromEuler(float rx, float ry, float rz) {
         // zyx
         Quaternion qx = Quaternion.getRoll(Vector3.X(), rx), qy = Quaternion.getRoll(Vector3.Y(), ry), qz = Quaternion.getRoll(Vector3.Z(), rz);
         
@@ -75,17 +75,17 @@ public class Quaternion implements Number<Quaternion> {
     }
     
     
-    public Quaternion roll(Vector3 axis, double rad) {
+    public Quaternion roll(Vector3 axis, float rad) {
         return roll(getRoll(axis, rad));
     }
     public Quaternion roll(Quaternion q) {
         return q.mul(this).mul(q.conj());
     }
-    public static Quaternion getRoll(Vector3 axis, double rad) {
-        return new Quaternion(Math.cos(rad/2), axis.norm().mul(1* Math.sin(rad/2)));
+    public static Quaternion getRoll(Vector3 axis, float rad) {
+        return new Quaternion((float)Math.cos(rad/2), axis.norm().mul(1* (float)Math.sin(rad/2)));
     }
-    public double abs() {
-        return Math.sqrt(w*w + x*x + y*y + z*z);
+    public float abs() {
+        return (float)Math.sqrt(w*w + x*x + y*y + z*z);
     }
     public Quaternion conj() {
         return new Quaternion(w, -x, -y, -z);
@@ -97,7 +97,7 @@ public class Quaternion implements Number<Quaternion> {
         return "[" + w + "," + x + "," + y + "," + z + "]";
     }
     
-    private static double sqrt(double x) { return Math.sqrt(x); }
-    private static double abs(double x) { return Math.abs(x); }
-    private static double pow(double a, double b) { return Math.pow(a, b); }
+    private static float sqrt(float x) { return (float)Math.sqrt(x); }
+    private static float abs(float x) { return (float)Math.abs(x); }
+    private static float pow(float a, float b) { return (float)Math.pow(a, b); }
 }
